Finalización de los indicadores
Los indicadores ya funcionan según los requisitos: Parsean la fórmula que los define y utilizan las cuentas o los indicadores nombrados por la misma. Falta implementar una interfaz para todo

+++ java/src/Cuenta.java
@@ -1,15 +1,15 @@
 class Cuenta {
-    private Short período;
+    private int período;
     private String nombre;
     private float valor;
 
-    Cuenta(Short período, String nombre, float valor) {
+    Cuenta(int período, String nombre, float valor) {
         this.período = período;
         this.nombre = nombre;
         this.valor = valor;
     }
 
-    Short getPeriod() { return período; }
+    int getPeriod() { return período; }
     String getNombre() { return nombre; }
     float getValor() { return valor; }
 }
